+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 **Added**
 
+- #1596 Allow to hide actions menu by using new marker interface IHideActionsMenu
 - #1588 Dynamic Analysis Specs: Lookup dynamic spec only when the specification is set
 - #1586 Allow to configure the variables for IDServer with an Adapter
 - #1584 Date (yymmdd) support in IDs generation
@@ -22,6 +23,7 @@ Changelog
 
 **Fixed**
 
+- #1596 Fix Reports page shows the Display/State/Add menu
 - #1595 Fix Wrong url in client's analyses profiles listing
 - #1593 Fix Out-of-range alert icon is shown to users w/o "View Results" privileges
 - #1592 Fix Publisher user cannot publish samples